+++ checklist.user.js
@@ -2,6 +2,7 @@
 // @name     Checklists
 // @version  1.0.0
 // @description Various changes to make checklist pages more readable
+// @include  https://ebird.org/checklist/*
 // @include  https://ebird.org/view/checklist/*
 // @include  https://ebird.org/*/view/checklist/*
 // @require  https://cdnjs.cloudflare.com/ajax/libs/jquery/3.3.1/jquery.js