+++ checklist_hide_age_table.user.js
@@ -2,6 +2,7 @@
 // @name     Checklist hide age table
 // @version  1.0.2
 // @description Hide the table showing the counts for age and sex
+// @include  https://ebird.org/checklist/*
 // @include  https://ebird.org/view/checklist/*
 // @include  https://ebird.org/*/view/checklist/*
 // @require  https://cdnjs.cloudflare.com/ajax/libs/jquery/3.3.1/jquery.js